HIVE-1040 [jira] use sed rather than diff for masking out noise in diff-based tests (Marek Sapota via John Sichi)
Summary:
Replace diff -I with regex masking in Java

The current diff -I approach has two problems:  (1) it does not allow resolution
finer than line-level, so it's impossible to mask out pattern occurrences within
a line, and (2) it produces unmasked files, so if you run diff on the command
line to compare the result .q.out with the checked-in file, you see the noise.

My suggestion is to first run sed to replace noise patterns with an
unlikely-to-occur string like ZYZZYZVA, and then diff the pre-masked files
without using any -I.

This would require a one-time hit to update all existing .q.out files so that
they would contain the pre-masked results.

Test Plan: EMPTY

Reviewers: JIRA, jsichi

Reviewed By: jsichi

CC: jsichi

Differential Revision: 597



git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1214011 13f79535-47bb-0310-9956-ffa450edef68