HIVE-2706 [jira] StackOverflowError when using custom UDF after adding archive after adding jars (Kevin Wilfong via Yongqiang He)
Summary:
https://issues.apache.org/jira/browse/HIVE-2706

The issue was that the current thread's classloader and the classloader in the
conf differed due to the prehook updating only the current thread's classloader
with new jars.  Now, it updates both classloaders, fixing the issue.

When a custom UDF is used in a query after add an archive, such as a zip file,
after adding jars, the XMLEncoder enters an infinite loop when serializing the
map reduce task, as part of sending it to be executed. This results in a stack
overflow error.

Test Plan: Verified it fixed the stack overflow error.

Reviewers: JIRA, heyongqiang, njain

Reviewed By: heyongqiang

CC: heyongqiang

Differential Revision: https://reviews.facebook.net/D1167

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1229885 13f79535-47bb-0310-9956-ffa450edef68