HIVE-2946 [jira] Hive metastore does not have any log messages while shutting itself down. (Vandana Kasavajjala via Ashutosh Chauhan)
Summary:
https://issues.apache.org/jira/browse/HIVE-2946

Addressed Ashutosh's comments.

When hive metastore is started, the event is logged with good amount of
information. But when it is stopped ( using a kill) command, no information is
being written to the logs. It will be good if we can add a shutdown event to the
HiveMetastore class.

Test Plan: EMPTY

Reviewers: JIRA, ashutoshc

Reviewed By: ashutoshc

Differential Revision: https://reviews.facebook.net/D2787

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1327323 13f79535-47bb-0310-9956-ffa450edef68