HIVE-3134 Drop table/index/database can result in orphaned locations (Kevin Wilfong via Carl Steinbach)
Summary:
Today when a managed table has a partition with a location which is not a subdirectory of the table's location, when the table is dropped the partition's data is not deleted from HDFS, resulting in an orphaned directory (the data exists but nothing points to it).

The same applies to dropping a database with cascade and a table has a location outside the database.

I think it is safe to assume managed tables/partitions own the directories they point to, so we should clean these up.

Test Plan: Added test cases to verify that the directories which would be orphaned are cleaned up by drop table/index/database.

Reviewers: JIRA, njain, cwsteinbach

Reviewed By: cwsteinbach

Differential Revision: https://reviews.facebook.net/D3633

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1351174 13f79535-47bb-0310-9956-ffa450edef68