CDH-31190 : (Tests portion only) HIVE-9743: Incorrect result set for vectorized left outer join (Matt McCline, reviewed by Vikram Dixit)
HIVE-9743 is a hairy backport and hence risky.
It needs other LLAP patches like HIVE-9937. Also, HIVE-9971, which is a cleanup patch, but can be worked around.

I confirmed that this is Tez-only patch by running the new tests introduced by this patch for MR.

Unless there are objections, we could bring in the test portion of this patch, which confirms the code
changes are Tez specific.

No conflicts.

    (cherry picked from commit 2531040758e796c5cc469a893c50a8f5a388ded6)