shuffle the metastore uris, so the request can be balanced to metastores
Currently, HiveMetaStoreClient connect to the first metastore uri defaultly If we have multi metastore uris. So the client only send request to the first metastore, the metastore will be the bottleneck to process the client request.

I add the logic to shuffle metastore uris, so that the request can be balanced to all of the metastores.