CDH-39701 : CLOUDERA-BUILD CDH-ONLY Fix for HIVE-9609 to avoid breaking API compatibility
The method getPartitions() was removed upstream in Hive 1.2 in HIVE-9609.
And replaced by getPartitionIterator() for memory efficiency.
Additionally, the new method fixes a bug for the case where
we can end up returning null when AddPartitionEvent was
initialized on a PartitionSpec rather than a List<Partition>.
Because we cannot change this public API in CDH5, we are
adding the bug fix in the old API.
When we move to CDH6, we no longer needed this CDH-ONLY fix
since at that point, it is ok to break API compatibility.
Sentry and possibly customer code extending MetaStoreEventListener
is vulnerable to this change.

Change-Id: I1f1de3c79483b3c694bf1223f228e29270db289c