HIVE-15711 :  Flaky TestEmbeddedThriftBinaryCLIService.testTaskStatus
The 'RUNNING' state can be set way before the job handle from MR frameworks is available to be set on the server in the taskDisplay, hence there is a possibility that we can receive jobStatus messages without the job id. Hence its not guaranteed  that it is  not null in running status hence removing this condition check. This will be checked anyways when the status is 'FINISHED'.Printing the status before validation so in case of validation failures we will know why it failed