CDH-54813 : CLOUDERA-BUILD : Introduce CDH version compatibility in CDHMetaStoreSchemaInfo
This change implements the isVersionCompatible method for CDHMetaStoreSchemaInfo. It makes sure that
value stored in SCHEMA_VERSION_V2 is compatible with the CDH version from the jars. Specifically,
it makes sure the SCHEMA_VERSION_V2 stored in the db is atleast at the minimum required version
so that HMS can safely start

Change-Id: I768e2c9cb6763058fdff119d63555a7448aeac67