Moved the other TypeInfos to serde-api
In order to move the other TypeInfo classes to serde-api we need to move the serdeConstants.java
as well. This is a thrift generated class. This commit copies the serde.thrift instead of moving.
The only reason I did not move it is in case of backwards compatibility reasons (in case someone is using the thrift file location to do something).
If it is okay to move serde.thrift from serde module to serde-api we can delete it from serde module in a separate
change.

The other concern is there are some TypeInfo classes which do some validation like VarCharTypeInfo, DecimalTypeInfo.
The validating methods use the actual type implementation like HiveChar, HiveDecimal etc to ensure that the params
are under the correct limits. This creates a problem since we cannot bring in the type implementations as well to
serde-api. Currently, I have marked these as TODO and commented them out.