Moved TypeInfoParser and TypeInfoFactory to serde-api
In order to use TypeInfoParser in standalone metastore, we should move it to serde-api
There is a problem in moving the TypeInfoParser to serde-api which is there are some validation util
methods which validated the max lengths of varchar, char or decimal typeinfo's precision and scale.
These values are originally defined in HiveChar, HiveVarChar and HiveDecimal which cannot be moved to
serde-api.

In order to fix this this change moved these constants to serde.thrift so that it can be accessed from
serdeConstants instead of HiveChar or HiveVarChar (not sure if this is the right thing to do, but there
doesn't seem any good option here)