Moved HiveChar, HiveVarChar, HiveCharWritable, HiveVarcharWritable to storage-api
The PrimitiveEntry caches (static maps) which keep a mapping between the typename to the classes needs to be
populated before using in standalone-metastore. This is done in StorageSchemaUtils static initialization block.
However, this initialization needs access to HiveChar, HiveVarchar and Timestamp related classes from serde.

I see that HiveDecimal class is in storage-api. HiveChar, HiveVarChar and their base classes are comparitively
light weight and can be moved to storage-api. However, Timestamp related primitive types like TimestampWritable,
TimestampLocalTZWritable, TimestampTZ, HiveIntervalYearMonth, HiveIntervalYearMonthWritable, HiveIntervalDayTimeWritable
classes cannot be moved in a straight-forward way because they have dependency with Java 8 and some util methods and
classes in LazyBinaryUtils. I have marked this as a TODO currently. As far as I can tell TypeInfoParser does not
need these classes while parsing type string to deserialize the types.

Note: HiveIntervalDayTime is already in storage-api. Not sure why the other time related types are not in storage-api
in the first place.