Fix for the tests
The issue is once we move PrimitiveEntry class to serde-api the internal tables which map
the string value to the respective type implementation classes are not populated due a race
condition. Previously the PrimitiveEntry was an inner class to PrimitiveObjectInspectorUtils
which caused the static initialization block to be called before we used PrimitiveEntry
and thus populating the type maps. When we move the PrimitiveEntry class outside, we need
an explicit mechanism to make sure that types are registered before we use PrimitiveEntry
util methods. In order to do that I introduced a Service called TypeRegistry which provides
the supported PrimitiveEntry objects which need to be registered. PrimitiveEntry static
initialization block determines in the runtime the needed TypeRegistry and registers
the PrimitiveEntry objects supported by the system