HIVE-20542: Insert NULL value for columns of NOTIFICATION_LOG for which values are not available
When no database is associated with an event we insert 'null' as database name in the metastore.
With this commit, we insert NULL as database name.

When no tablename is associated with an event we insert an empty string as table name in the
metastore. With this commit, we insert NULL as table name.

Even if a catalog name is associated with an event, addNotificationLog() doesn't insert catalog in
the metastore. With this commit we take care of that as well.

Ashutosh Bapat.