HIVE-20644: Additional changes per Thejas's following comment
For completeness, maybe make similar update in
ql/src/java/org/apache/hadoop/hive/ql/exec/spark/SparkReduceRecordHandler.java as well.

The change in the previous commit to the above file is required without any doubt since it exposes
the row. But the changes in this commit are a bit dubious since I am not sure if the key or value
contains any sensitive information.

Ashutosh Bapat