HIVE-20708: Load an external table as an external table on target with the same location as on the source
Dump an external table as an external table.

When loading an external table set the location of the target table same as the location of source,
but relative to the file system of the target location. IOW, the scheme, authority of the target
location is same as the target file system but the path relative to the file system is same as the
source.