HIVE-20953: Remove a function from registry if adding it to metastore fails.
A function being created is added to function registry and then added to the metastore. If the later
fails, we do not clean it from the function registry. Thus that function remains available to that
session. This commit fixes that by removing a function from registry if addition to metastore fails.