HIVE-21078: Replicate column and table level statistics for unpartitioned Hive tables
The column statistics is included as part of the Table object during bootstrap dump and loaded when
corresponding table is created on replica.

During incremental dump and load, UpdateTableColStats event is used to replicate the statistics.

In both the cases, the statistics is replicated only when the data is replicated.

Ashutosh Bapat