HIVE-21078: Fix for test failures in ptest run 15625
When an UpdateTableColStat messages is processed for dump the underlying table may have been
dropped. In this case, Utils.shouldReplicate() checks will fail with "no table found" error. But we
require the Table object for checks in Utils.shouldReplicate(). Hence include the table whose column
statistics is being updated in the event message.

Ashutosh Bapat.