HIVE-21109 : Support stats replication for ACID tables.
During bootstrap we use a method similar to non-ACID tables to transfer statistics of an ACID table
from source to replica. However installing statistics of an ACID table requires a valid writeId and
writeId list. We use the table/partition's latest writeId and a valid transaction list containing
only that writeId to install the statistics in the metastore.

During incremental replication writeId is obtained from the UpdateStats event and valid writeId list
with that writeId marked as valid is used to install the column statistics. Table level statistics is
replicated by replaying corresponding ALTER_TABLE/ALTER_PARTITION event.

Further this commit has following related changes.

1. The table or the partition associated with the commit transaction event should have been created
when replaying corresponding events before commit transaction event. Thus there is no need to add
tasks for creating the table or the partition.

2. Use a cooked up valid write Id list with given writeId for replication
When passing valid writeId list to metastore during replication, pass a valid write Id list which
contains the given writeId as valid write Id in it. During replication, there can be many
transactions open at a time and not all of them are visible to each other. Similarly for
corresponding writeIds. Thus the table snapshot may not contain the given writeId, under which we
are performing an operation, as valid. So instead we cook up a valid writeId list with given writeId
as a valid writeId.