HIVE-21683. Fix FileSystem shims for Hadoop 3.3
Hadoop 3.3 added a new openFile() call in HADOOP-15229. This adds the
appropriate shim to ProxyFileSystem23.

Note that this can't be committed to the upstream Hive project until
Hadoop 3.3 is released and Hive upgrades its dependency. HADOOP-15229
was backported to CDPD, so we need to commit this to CDPD as well to
unbreak tests.

Testing: Verified that I could run a qtest after this change. Without
the patch, all qtests failed with "Wrong fs: pfile:///..." errors.

Change-Id: I946843e3c3953046bcc8940e7cf5cd01514b7d2b