HIVE-20983: Reduce memory requirement for hash map textExpand tests
The tests use an unrealistic load factor to make sure the hashmaps expand at every put. Since this improvement, hashmaps expand to 4096 buckets at first put, so overall the tests were using too much memory. Now the tests are limited to 6 expansions instead of the original 18. This allows the tests to expand up to 2^18 buckets as they did in the original test case.