HIVE-21975: Fixed incremental compilation.
When compiling files one by one, a compile error was produced
when assigning a Task<? extends Serializable> to a Task<?>.
This checkin fixes the compile error (related to HIVE-20008).

Sidenote: Even though most of the calls are more generalized,
the "extends Serialiable" clause is left in a few key places
to ensure that the Task object is always instantiated with a
serializable parameter.