HIVE-22338: Shade dependent jars into the kudu-handler
This patch shades the Kudu client and it’s async dependency into the
kudu-handler jar to simplify add jar style usage and ensure the
addDependencyJars method adds all the required classes.

It also adjusts the master address property to fallback to the default
instead of looking it up first. This allows the default to be unset when
the table property exists. I ran into this when using the
handler via `add jar`.

Change-Id: Ie900abe21b685c9591e6147efe5c3b4b96791f5e