HIVE-22537: getAcidState() not saving directory causes multiple filesystem calls
getAcidState() can make use of single recursive directory listing call and save a snapshot of the directory. Later this snapshot is used to probe existence of some files, such as _acid_version. This optimization was turned of in OrcInputFormat. Even after turning this on, isCompactedBase and isMetadataFile were not using the snapshot saved.

Change-Id: I9eb69b1715c7f577ec5e5b21e505bbed5a953230