CDPD-6860: Implement mapping of Calcite functions to Impala
This checkin adds the resource files that contain Impala function signatures.
Calcite rules are added that take the Calcite function names and types and
maps them onto Impala functions and types.

The resource file contains both the Impala name and Calcite name (e.g.
Impala's "multiply" is Calcite's "*").

There are cases that are handled in Hive where we need to cast operand types
to get a match in Impala.  For instance, Impala does not have a "sum(INT)"
method, so if the operand is INT, we need to add a "sum(CAST(int_column AS BIGINT))"
function.

This checkin will compile, but will not work until the rest of the FENG framework
is added.

Change-Id: I13a857ad892dfb7806d4a781f535d2a57e9b85a7