CDPD-8178: More funcmapper changes for FENG
After committing CDPD-6860, there were some more additions and fixes
needed in this module once the caller was implemented.

Included in this are:

- RelNodes need to be created from HiveRelBuilder, not the Calcite one.
- Scalar and Agg FunctionDetail similarities are combined.
- Comparing SqlTypeNames when needed as opposed to RelDataTypes.
- More utility functions added to ImpalaTypeConverter
- resource file fixes and additions
- Utility class to create Impala ScalarFunctions

In addition to this, a hack was added to support constant chars being
compared with varchars.  A separate Jira (CDPD-8173) was created to
make sure this works when comparing non-constant chars to varchars and
to perhaps fix this in a different way.

Change-Id: Ic4ff1c444ff0740fcae5bbf4724787837f33f9cd