CDPD-8179: Propagate Impala ExecRequest through to ImpalaCompiler
This includes all changes to support using the Impala framework.
The main portion of these changes are in CalcitePlanner and SemanticAnalyzer.
From there, it will call out all Impala specific methods.

While this code is Impala specific, it has been written so that we should
be able to place all Impala specific modules into a "plugin" module with
very little change, should we decide to do so.

Change-Id: I6ef97d40bc6f3915336be81b517a0497315c1f91