CDPD-8436: Better error message when function is not found
Before this checkin, A NPE would be thrown if a function was
not found in the resource file.

Note though that this condition should never happen in production,
since all Calcite names in the resource file should map to an Impala
function.  But this error message will help us in the debugging stage.

Change-Id: I7b980b9a18b4dc410c0564f9eed29afc2a23513b