CDPD-8694: Fix Impala Case handling
The case function signature comes from the Impala function
which is of the form:
CASE [expr] WHEN expr THEN expr [WHEN whenexpr THEN rettype ...] [ELSE elseexpr] END

In Calcite, the "WHEN" expression is always of type <boolean> and the THEN expression
is always the same type as the return type.

In the case where the optional first expression (i.e. CASE [expr]) is given, Calcite rewrites
the WHEN expression so that all WHEN expressions are of type <boolean>
(e.g.WHEN expr == whenexpr)

The ELSE condition in Calcite will also always be the same type as the return type.

Impala simplified this signature for the function to be case(<rettype>) since all other
arguments are optional or derived.

The purpose of this commit is to help map the Calcite signature at runtime to the Impala
signature.

Change-Id: I889fc894a0c0f5a6afd4d94c19e99ced30463127