CDPD-8824: Enable NoDag option for Impala Calcite Rules
The NoDag=true option allows for separate equivalent nodes to be
created.

This is necessary to do when converting Calcite RelNodes to Impala
RelNodes. There was a case when a duplicate table/ScanNode is read
within the query.  Calcite treats these RelNodes as equivalent, but
Impala needs these to be distinct RelNodes.

It also helped in the case where we had similar RexNodes.  Impala
needs some operands to be the same type (e.g. =(BIGINT, BIGINT),
not =(BIGINT, INTEGER)).  There was a case where the second operand
needed a cast of the RexLiteral(0) from INTEGER to BIGINT, but these
RexNodes were equivalent which made the RelNodes equivalent.  The
NoDag solves the problem in this case as well.

Change-Id: Ic5314bedecb98e94e4a06c6c8ddd3f07c6a58dab