CDPD-8693: Fix Impala extract function handling
The time specific extract functions in Calcite are of the form
"EXTRACT(FLAG(YEAR):SYMBOL, DATE)" while the Impala signatures
are of the form "YEAR(DATE)". When doing the mapping, the first
parameter needs to be removed.

This applies to other "extract" functions as well (e.g. month).
This commit will not handle the general "extract" function for
which a jira, CDPD-8867, has been filed.

Change-Id: Iffbcbb514b3f1398cec8cbb2eea61c8773c3d275