CDPD-8540: "in" predicate not working
Needed to change the resource file so the "in" Calcite function
was pointing to an Impala method.

Currently it is set up to point to in_set_lookup.  This works for
TPCH, but more research needs to be done to see if other "in" functions
need to be supported. CDPD-8877 has been created for this task.

Change-Id: I75552435759795e6c9a51c709a94478de2238cb2