HIVE-21939: Bump protoc version in standalone metadata
Standalone metadata package currently depends on proroc 2.5.0 which does not support Aarch64 platform and blocks building Hive on Aarch64. This patch proposed to bump  protoc used in standalone metadata to 2.6.1, and also added a new profile that detacts hardware architecture, override the default protoc package with package from com.github.os72(it has aarch64 support) if the hardware arch is Aarch64. By doing this, we will keep the affects to current X86 users to the minimum.