CDPD-9131: Fix explain terms in Impala RelNodes
A bug in the "explain terms" code was exposed while running through the
integrated tests.

The Impala "explainTerms" method was calling the associated Calcite RelNode's
"explainTerms" method, and that method called the "getInput" method, which
caused issues.

The solution is to copy most of the Calcite "explainTerms" method into the
Impala code.

Change-Id: Ibbaa77a2821b53483096969c16483cd95be0b25d