CDPD-7874: Added support for TIMESTAMP
This checkin provides basic support for Calcite TIMESTAMP literals.

This is not done in the most optimal way. The TIMESTAMP is converted
to an epoch time and then cast to a TIMESTAMP.

The optimal way of handling TIMESTAMP literals is to do constant folding.
This is not yet supported at this commit time. CDPD-9411 was created to
track this followup.

Even though this commit isn't the optimal way, it is still needed in the
case where constant folding isn't supported.

Change-Id: I25851e66071ed7238e975f4cb1feb6844395a611