CDPD-9581: Power function not mapping correctly
A decimal argument into the power function should be cast into
a double, since the Impala power function only takes a double
as an argument.

Note that this can apply to functions other than power. It will
apply to any function that only has mappings for type 'double'.
Also note that if a function has both decimal and double mappings,
the function with the matching return type will be selected.

In addition to this change, we want to eliminate the mapping of
functions that have different return types. This checkin will
ensure that this type of function will not be selected, and as a
bonus, it simplifies the code a little.

Change-Id: Id8403cd17d2154e72d988345ad5aef166d37b749