CDPD-9372: Fix handling of sq_count_check() and related changes to Sort rule
Added sq_count_check() to the resource file.

On top of this change, a "Filter<-Sort" rule needed to be created. To handle
this, the HiveSort -> Impala transformation is now broken up into 2 rules:

1) Filter <- Sort<- any()
2) Sort <- any()

The "any()" has 2 paths, one where the node is a "Project" and one where it
is anything else.

Change-Id: Idf955116702792ac2159bcfa727c2449c27b5d52