CDPD-9067: Support proper Impala session scope and parameters
This change allows an ImpalaSession to be used across multiple statements
and supports setting of impala config options.

Session scope is handled by ImpalaSessionManager which is implemented
similar to SparkSessionManager. The ImpalaSession is established on first
use and associated with the HiveServer Session. The ImpalaSession may be
closed and reopened if any Impala conf options are changed.

Options are supported by adding any HiveConf setting with an "impala"
prefix to the HS2 OpenSession request config. Impala backend will overlay
these option values onto its session parameters and log warnings for any
invalid settings.

Testing:
Manual testing with Impala backend. Confirmed with debugger that Impala
options are forwared to the backend and that changes cause sessions to be
reopened. Tested backend restart with open session.

Change-Id: I27e14a9aea76c428484326610bd8befcc66c1453