CDPD-12963: Use interpretation on RexLiteral to determine Impala type
The datatype for a RexLiteral isn't determined directly by the rowType.
Instead, it has to be obtained through the HiveNlsString which has an
"interpretation" which is char, varchar, or string.

Couple of things had to change in order to handle this.

- Casting is ignored for char/varchar/string types.  It is assumed that
the caller will already have arranged for compatible types.

- Before this checkin, the Calcite class used to determined the datatype was
SqlTypeName. This proved inadequate because there is no differentiator
between varchar and string. So instead, a default RelDataType is being
used to determine the Calcite class.

- The ImpalaTypeConverter class was cleaned up to isolate PrimitiveTypes
within the utility class.

Change-Id: I17604426329438f6b1f6178d22a5de5192c4243c