CDPD-8690: Allow canonicalization of equijoin predicates with arbitrary expressions
Previously, equijoin predicates of the form a = b were canonicalized
only if each side was a RexInputRef. This patch supports doing this
for arbitrary expressions in the equijoin predicate. Tested with
TPC-DS q2, q51, q59 which were problematic earlier. Also tested
several TPC-H queries.

Additionally, the ImpalaUnionRel was not setting the SlotRef labels
which caused the Explain output to show null values. This patch fixes
it.

Change-Id: I259c644fb1fb13a909c0a376b643b8eb6a7a753e