CDPD-11977: Support constant folding of expressions
This commit now enables constant folding.

There are no qtests for this because it requires libfesupport.so
to be present. If the library is not present, the query will still
work, but the expression will not be folded.

A little rewiring had to be done because the ImpalaBasicAnalzyer
needed to be created earlier in the compilation.

Change-Id: I513b224ccf2ebdebd0e3cdc63d113648447b7462