CDPD-13191: In clause failing when all impala types are CHAR
There are some functions (like "in") where Impala does not have handle
CHAR and VARCHAR datatypes but does handle STRING datatypes. These
datatypes need to be converted to find the correct signature. These
datatypes also need to be cast to STRING in the operands and need to
be cast back to the original type after the function is called.

The preferable way to handle this would have been to do the casting up
front when we call the ImpalaFunctionResolver to resolve the datatypes.
However, the char-like datatypes that Hive passes into the resolver are
not always the final datatype passed into the pipeline. So the casting
cannot be done at this point.

Change-Id: I375cda7ade98b334f147d61ae2b8132fc57a0766