HIVE-23965: Improve plan regression tests using TPCDS30TB metastore dump and custom configs
1. Add new perf driver, TestTezTPCDS30TBCliDriver, relying on a dockerized metastore.
2. Use Dockerized postgres metastore with TPC-DS 30TB dump
3. Remove old drivers (with and without constraints), related classes
(e.g., MetastoreDumpUtility), and resources.

After discussion in the JIRA we decided that keeping the old drivers
is most likely useless.

4. Use Hive config properties obtained and curated from real-life usages
5. Allow AbstractCliConfig to override metastore DB type
6. Rework CorePerfCliDriver to allow pre-initialized metastores

Remove system property settings in the initialization of the driver and
leave in the configuration to set it up if needed. This is necessary to
be able to use the driver with a preinitialised metastore.

Remove redundant logs in System.err. Logging and throwing an exception
is an anti-pattern.

Replace assertions with exceptions and improve the messages.

7. Upgrade postgres JDBC driver to version 42.2.14 to be compatible
with the docker image used
8. Update path to new postgres version in TestBeelineArgParsing
9. Display plans using hive.explain.user set to false which seems
to be preferrable for performing a diff and checking regressions

10. Disable queries 14 (HIVE-24167), 30 (HIVE-23964)
11. Re-enable CBO plan tests for queries 44, 45, 67, 70, 86

The queries were disabled as part of HIVE-20718. They were supposed to
be fixed in Calcite 1.18.0 and currently Hive is in 1.21.0 so it is not
surprising that they pass.

12. Add missing queries: cbo_query41, cbo_query62, query62