CDPD-13616: Use Extended mode for query profile
Use the EXTENDED mode for the explain output unless it is an
EXPLAIN statement. This means for the query profile, the explain
output will be EXTENDED mode, which matches Impala's output.
For an EXPLAIN statement, we will continue to use the query
options's explain_level.

Testing: Ran queries and verified query profile. Ran Explain
with different explain_levels and verified expected output.

Change-Id: Ib4ca24a446346912b5a3d1079d0a466a458f1a10