CDPD-10642: Add test for Catalog's get_partitions_by_expr
This change adds a unit test for Catalog service's HMS
get_partition_by_expr API. The test assumes that a catalog service
is running with --start_hms_service=true flag. By default it assumes
that the catalog is running on localhost:5899 but this can be
overridden. Also, Catalog service should have the hive-exec.jar
in its classpath for this test to work.

Change-Id: Ib9e51001edc7b98045f71e6bc80208163ab4e187