CDPD-13192: Fix mismatch of expressions for the aggregate node.
When Impala creates the slot descriptors for the aggregate node, it may
not necessarily directly match the order given by the ImpalaAggregateRel.

This commit fixes the mapping so that the index of an InputRef above
the Aggregate Node points to the correct SlotRef.

All the SlotRefs are created within the Impala AggregateInfo object
except for groupingId. If there is a grouping id column, an extra
SlotRef is created in the ImpalaAggregateRel object.

Change-Id: I01689f34c9e0f73d090276f67445537b827c77e0