HIVE-23780: Fail dropTable if acid cleanup fails
Otherwise, metastore is left in inconcistent state should acid cleanup fail after table was dropped.

Existing test infrastructure uses a hack to create transactional tables of metastore on demand. These tables are created when txnhandler is invoked the first time. For many tests, this happen in a dropDatabase call. This leads to a deadlock because dropDatabase call now holds lock on database record when on demand initialization requires the same lock. Therefore, this commit also removes this hack and introduces proper initialization of transactional tables to all the tests that require it. This on demand initialization happens only in testing, therefore there is no risk introduced by this change to production system.

Change-Id: Ica7666afe40cb0f0128266c9c3f6ebc560b24c0e