CDPD-13352: Request the file metadata for pruned partitions only
A centralized ImpalaTableLoader is added in this commit.
Its purpose is to manage the loading and caching of
table and partition metadata of all tables (across all
query blocks) within a query. It gathers the list
(de-duplicated) of partitions needed for a particular
table. Then invokes the load request for those partitions
from Impala.

This patch does not guarantee that only those partitions
will be eventually loaded by Impala. That may require
additional changes on Impala side.

Change-Id: I7f5d5b0aa5381d9d783195f88c5039d4f84e2a79