CDPD-12463: Enable computing Impala statistics from FENG
This patch generates an Impala COMPUTE STATS query from an
ANALYZE TABLE statement, i.e., it executes stats computation
by sending a SQL statement to the coordinator instead of an
optimized plan.
Thus, it relies completely on Impala to generate and publish
the stats accordingly.

This commit does not contain secrets

Change-Id: I12521536a5f0547ddb425d7ea355a0bda29cc66a