CDPD-15044 Cast to string implemented properly.
Whenever a datatype was cast to a string, the compilation failed because
it tried to convert the word "STRING" into an SqlTypeName and this did not
exist.

With this checkin, it will create the normalized Calcite datatype based on
the casted Impala type.

Also, if the return type is already known at the time the function resolver is
called, that Calcite type will be used.

Change-Id: Ib9fbeca135237071d28e415f40bf3ab5cd06e1f6