CDPD-14844: Retrieve executor membership snapshot from Impalad
This patch retrieves the executor membership snapshot using the
thrift interface added to ImpalaServer as part of CDPD-15086.
We then use it to set the ExecutorMembershipSnapshot static
instance which is initialized once when the first query is
submitted after HS2 restart. A follow-up patch will allow timer
based caching of the executor membership.

Tested the changes by examining explain plans for a few TPC-DS
queries..they now correctly show instances=3 instead of the default
20 that it was showing previously.

Change-Id: I97e2d3f00ccfedae183c1efc0ae14ba7d77e41b4