HIVE-21196: Update plans in Perf cli drivers
With the new estimation in the benefit of the bloom the multi-col
semijoin reduction kicks in also in query5.

Interestingly there are also some important changes in query24. Due to
the new estimation a multi-col semijoin reduction on the same columns is
used in two scan operators (on the same table). This also triggers the
shared work optimizer that is able to merge the semijoin branches and
the respective scan operators.