FENG-15947: Incorrect function picked for 'between'
For tpch queries 7 and 8, the "string between(boolean, string,
string, string) function should have been picked, but instead
it was picking "date between(boolean, date, date, date)".

The reason for this was because of the way the functions are
stored in the HashMap, but requires a bit of an explanation.

The HashMap holding these functions has a key of "ImpalaFunctionSignature"
which is of "Comparable" type.  The hashCode calculation for
between only works off of the first argument for the signature
(if the first argument exists), and all of the prototypes for
"between" have a boolean as its first argument.  So all of the
"between" functions are mapped to the same bucket.

When HashMap detects that the key is "Comparable", it uses TreeNodes
within the bucket to find the key.  So far, so good...

The returnType is not used when looking for the signature in the
HashMap.  This is because at the point of retrieval, the returnType
may need to be calculated and thus provided as null.

So this was the cause of the problem.  The "compareTo" function did
look at the returnType in processing.  When creating the tree, the
returnType always had a value (it reads the signatures from a json
file).  When traversing the tree for a match, the signature was
"null" and thus sometimes went down the wrong path of the tree,
depending on how the tree was set up.

Since we only really need to do comparing when the function names
are the same (for choosing a function for cast checking), the
comparing portion is now run as a Comparator in the specific place
where it is needed.

Change-Id: I2545cf7b77530acdb988073a312e6845fedc8197